+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2023-12-30
+
+### Added
+
+- Warning for deprecated tooth.json format versions.
+- `info.source` field for indicating the source repo of a tooth.
+
+### Changed
+
+- Check tooth repo validity with Go official module library.
+
+### Fixed
+
+- Failed to uninstall a tooth when some of its files are already removed.
+
 ## [0.16.1] - 2023-10-11
 
 ### Fixed
@@ -304,7 +319,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/lippkg/lip/compare/v0.16.1...HEAD
+[unreleased]: https://github.com/lippkg/lip/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/lippkg/lip/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/lippkg/lip/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/lippkg/lip/compare/v0.15.2...v0.16.0
 [0.15.2]: https://github.com/lippkg/lip/compare/v0.15.1...v0.15.2